LayoutTests/inspector/debugger/disable-script-expected.txt

Tests that inspector effectively disables script execution in the inspected page.

Execution count: 2

Running: init
Script execution status: allowed

Running: tryWithEnabled
Attempting execution...
Script execution status: allowed

Running: tryWithDisabled
Attempting execution...
Script execution status: disabled

Running: tryWithRestored
Attempting execution...
Script execution status: allowed

Running: finalize
